put creation of api in new spot
